Mainクラスにべた書きしていた処理を、適切なレイヤーに分割しました。これらに対して、テストを実装します。 #18
