+++ src/main/java/me/u6k/plantuml_image_generator/model/WebRepository.java
@@ -1,0 +1,8 @@
+
+package me.u6k.plantuml_image_generator.model;
+
+public interface WebRepository {
+
+    String getUrl(String url);
+
+}
